move mscw ctm reader helpers to plugin
